search: Add API to search invisible text
The underlying GtkTextView has this capability, so we can expose an API
in GtkSourceSearchSettings to do this. The API is a property named
'visible-only' and a setter and getter function. The property defaults
to TRUE, giving the previously existing behaviour. If you set the
property to false, then search matches may include invisible text.

Closes: #51