gutterrendererlines: avoid PangoLayout for line numbers
This can save a huge amount of repetitive work in PangoLayout and
snapshoting a layout. Instead, it caches Pango glyph information and
uses that to build a custom PangoGlyphString for GskTextNode.

This is similar in spirit to the optimization work for Pango rendering
that I put into VTE this cycle.