view: Fix Tab key with IBus pre-edit under Xorg
When Tab key is pressed with input-method pre-edit mode in ibus-anthy
or ibus-typing-booster, the key can be a control key of the candidate
popup window and it needs to be handled before GtkSourceView handles
the auto-indent features. Otherwise the pre-edit and the candidate
popup window are shifted with the Tab width.

Now GtkSourceView connects to "insert-text" signal in GtkTextBuffer
to get "commit" signal in GtkIMContext and handle the auto-indent
if the inserted text is same with the typed keyval.

Fixes: #320