view: reset current line style properties after unsetting the style scheme
When the `style-scheme` property is unset (i.e. NULL), GtkSourceView was
keeping the previous style (the "classic" one by default) for rendering the
current line, also in the gutters.

Make sure the current line style properties used by the view and gutters
are properly reset after unsetting the style scheme.