map: dont ensure_highlight() from snapshot
This has to be done in GtkSourceView just given how things work. But the
GtkSourceMap generally has a lot in view and we don't want to try to do
the ensure_highlight() here beacuse we're too late in the frame clock
cycle for that sort of non-deterministic time.

We already do it from size_allocate() which should be okay given that we
are generally scrolled. The visible area for the main view should pick up
the rest of the changes for us.