indenter: keep indent for cursor at line start
If the cursor as at the beginning of the line, avoid performing
default indentation. Either there is text after the cursor that
may be maintained or there is nothing after the cursor and at
column 0 which would have nothing to copy anyway.

Discovered and fixed by Y Somda gitlab.gnome.org/yoeo_s

Fixes: #366

Signed-off-by: Christian Hergert <chergert@redhat.com>