iter: add short-circuit for checking inside word
If we have alnum text and so is the previous (or there is no previous)
character, then we can short-circuit here and get the same answer.

This can help considerably on double clicking text to extend the selection
when the word is a minified JSON document. There is still more to do, but
this halves that work.

Related: GNOME/gnome-text-editor#742