utils: raise resource:// priority
Now that we bundle our internal language specs (both to reduce IO as well
as be certain they are available) we want those to take priority over other
installations.

For exapmle, if you install to ~/.jhbuild or something like that, then the
system install (which is an older version) may take precidence over what
we have bundled.

Instead, only allow overriding the internal versions in the user data dir
or the same installation as the library. Otherwise, they are prioritized
below the resource bundled version.

This fixes an issue where even with color updates in Adwaita.xml, they are
not getting picked up because the older system installation.

Related: !376