view: draw current line highlight for single-line selection
This is what the comment actually says we are doing, which it wasn't. This
is a bit of a behavior change from previous, but necessary to be able to
have more control over it from applications.

To keep our other behavior, an application can disable the setting when
the focus leaves the widget. But you cannot do the opposite without this
change.

Related: GNOME/gnome-text-editor#710