gtksourceview: Add typecast when setting source_buffer
With the GNOME 48 runtime, build fails with:

    gtksourceview.c: In function ‘set_source_buffer’:
    gtksourceview.c:1589:43: error: assignment to ‘GtkSourceBuffer *’ {aka ‘struct _GtkSourceBuffer *’} from incompatible pointer type ‘GtkTextBuffer *’ {aka ‘struct _GtkTextBuffer *’} [-Wincompatible-pointer-types]
     1589 |                 view->priv->source_buffer = g_object_ref (buffer);
          |                                           ^

I believe this is due to newer GLib versions preserving type information
in g_object_ref(). This code path is guarded with a GTK_IS_SOURCE_BUFFER
(buffer) check so the assignment is correct. Add a cast.

(On newer branches, this was fixed in commit
aa3140aa69760b62e4aaea726801951c1cb6b8b5 but that does not apply cleanly
to the gnome-3-24 branch.)